requirements.txt
matplotlib==3.4.2
numpy==1.20.2
htbuilder==0.5.0
Pillow==8.3.1
streamlit==0.84.1
tensorflow-cpu==2.8.0rc1

